Purges deprecated patient list component
