+++ packages/esm-commons-lib/translations/en.json
@@ -13,7 +13,6 @@
   "errorCopy": "Sorry, there was a problem displaying this information. You can try to reload this page, or contact the site administrator and quote the error code above.",
   "gender": "Gender",
   "hivPositive": "HIV Positive",
-  "lastVisit": "Last Visit",
   "moveToListSideNav": "Move to list",
   "name": "Name",
   "new": "New",
